Create PersonLink component with functionality to highlight links
