+++ src/components/PeoplePage/PeoplePage.tsx
@@ -1,73 +1,14 @@
-import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { Loader } from '../Loader';
-import { Person } from '../../types';
-import { getPeople } from '../../api';
 import { PeopleTable } from '../PeopleTable';
 
 export const PeoplePage = () => {
-  const [people, setPeople] = useState<Person[]>([]);
-  const [error, setError] = useState<boolean>(false);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
   const { personSlug } = useParams();
-
-  useEffect(() => {
-    getPeople()
-      .then(allPeople => setPeople(allPeople))
-      .catch(() => {
-        setError(true);
-      })
-      .finally(() => setIsLoading(false));
-  }, []);
 
   return (
     <>
       <h1 className="title">People Page</h1>
       <div className="block">
-        <div className="box table-container">
-          {isLoading && <Loader />}
-
-          {error && (
-            <p data-cy="peopleLoadingError" className="has-text-danger">
-              Something went wrong
-            </p>
-          )}
-
-          {(people.length === 0 && !isLoading) && (
-            <p data-cy="noPeopleMessage">
-              There are no people on the server
-            </p>
-          )}
-
-          {!isLoading && (
-            <table
-              data-cy="peopleTable"
-              className="table is-striped is-hoverable is-narrow is-fullwidth"
-            >
-              <thead>
-                <tr>
-                  <th>Name</th>
-                  <th>Sex</th>
-                  <th>Born</th>
-                  <th>Died</th>
-                  <th>Mother</th>
-                  <th>Father</th>
-                </tr>
-              </thead>
-
-              <tbody>
-                {people.map(person => (
-                  <PeopleTable
-                    key={person.slug}
-                    people={people}
-                    person={person}
-                    personSlug={personSlug}
-                  />
-                ))}
-              </tbody>
-            </table>
-          )}
-        </div>
+        <PeopleTable personSlug={personSlug} />
       </div>
 
     </>
