+++ src/components/PeopleTable/PeopleTable.types.tsx
@@ -1,7 +1,3 @@
-import { Person } from '../../types';
-
 export interface PeopleTableProps {
-  people: Person[];
-  person: Person;
   personSlug: string | undefined;
 }
