+++ src/components/PersonLink/PersonLink.types.ts
@@ -1,0 +1,7 @@
+import { Person } from '../../types';
+
+export interface PersonLinkProps {
+  people: Person[];
+  person: Person;
+  personSlug: string | undefined;
+}
